[add]レスポンシブ対応毎に（mentor用の質問コンポーネント）をコンポーネント化
