+++ resources/js/components/Mentor/Question/Show/question.js
@@ -1,43 +1,15 @@
 import React from "react";
-import Typography from "@material-ui/core/Typography";
-import Box from "@mui/material/Box";
-import Avatar from "@material-ui/core/Avatar";
-import SelectStatus from "../../../Atom/Select/SelectStatus";
 import BreakingPoint from "../../../BreakingPoint";
 import useMedia from "use-media";
-import { Link } from "react-router-dom";
-import Button from "@mui/material/Button";
+import QuestionContentMobile from "./responsive/questionContentMobile";
+import QuestionContentPc from "./responsive/questionContentPc";
 
 /**
  * 質問表示
  */
 function Question(props) {
     const isWide = useMedia({ minWidth: `${BreakingPoint}px` });
-    const categories = ["カリキュラム", "成果物"];
-    const topics = [
-        // カリキュラムのトピック
-        "AWS",
-        "HTML",
-        "CSS",
-        "JavaScript",
-        "サーバー",
-        "PHP",
-        "Laravel",
-        "データベース",
-        "Git&GitHub",
-        "マイグレーション",
-        "リレーション",
-        "認証・認可機能(カリキュラム)",
-        "API(カリキュラム)",
-        "その他(カリキュラム)",
-        // 成果物のトピック
-        "認証・認可機能(成果物)",
-        "API(成果物)",
-        "画像処理",
-        "Heroku環境",
-        "デザイン",
-        "その他(成果物)"
-    ];
+
     let question = [];
     if (props.question) {
         // 最後の文が認識されないのを防止する
@@ -190,240 +162,34 @@ function Question(props) {
         <>
             {isWide ? (
                 // 大画面で表示するコンポーネント
-                <Box sx={{ width: "90%", marginX: "5%" }}>
-                    <Box
-                        sx={{
-                            display: "flex",
-                            alignItems: "center",
-                            justifyContent: "space-between"
-                        }}
-                    >
-                        <Box sx={{ display: "flex", alignItems: "center" }}>
-                            <Avatar
-                                alt="Student"
-                                src="/images/pose_english_shrug_man.png"
-                                sx={{ marginRight: 2 }}
-                            />
-
-                            <Typography variant="h7" component="div">
-                                受講生 &nbsp; {props.updated_at}
-                            </Typography>
-                        </Box>
-
-                        <Typography variant="h7" component="div">
-                            {categories[props.category]}&nbsp; /&nbsp;
-                            {topics[props.topic]}
-                            &nbsp; /&nbsp;
-                            {props.curriculum_number}&nbsp;
-                        </Typography>
-                    </Box>
-
-                    <Box sx={{ display: "fex", position: "relative" }}>
-                        <Typography
-                            variant="h4"
-                            component="div"
-                            sx={{
-                                width: "70%",
-                                marginTop: 4,
-                                fontSize: 30,
-                                fontWeight: "bold"
-                            }}
-                        >
-                            {props.title}
-                        </Typography>
-                        <Box sx={{ position: "absolute", right: 0, margin: 2 }}>
-                            <SelectStatus
-                                id={props.id}
-                                status={props.status}
-                                setStatus={props.setStatus}
-                                responseStatus={props.responseStatus}
-                            />
-                        </Box>
-                    </Box>
-
-                    <Typography
-                        variant="h6"
-                        component="div"
-                        sx={{
-                            marginTop: 3,
-                            borderBottom: "1px solid gray",
-                            fontWeight: "bold",
-                            padding: 1
-                        }}
-                    >
-                        調べたこと
-                    </Typography>
-                    <Typography
-                        variant="h5"
-                        component="div"
-                        sx={{
-                            marginTop: 1,
-                            padding: 2,
-                            marginBottom: 2
-                        }}
-                    >
-                        {props.remarks}
-                    </Typography>
-                    <Typography
-                        variant="h6"
-                        component="div"
-                        sx={{
-                            marginTop: 4,
-                            borderBottom: "1px solid gray",
-                            fontWeight: "bold",
-                            padding: 1
-                        }}
-                    >
-                        試したこと、分からないこと
-                    </Typography>
-                    <Typography
-                        variant="h5"
-                        component="div"
-                        sx={{
-                            marginTop: 1,
-                            padding: 2,
-                            marginBottom: 2
-                        }}
-                    >
-                        {question}
-                    </Typography>
-                </Box>
+                <QuestionContentPc
+                    updated_at={props.updated_at}
+                    category={props.category}
+                    topic={props.topic}
+                    curriculum_number={props.curriculum_number}
+                    title={props.title}
+                    id={props.id}
+                    status={props.status}
+                    setStatus={props.setStatus}
+                    responseStatus={props.responseStatus}
+                    remarks={props.remarks}
+                    question={question}
+                />
             ) : (
                 // スマホで表示するコンポーネント
-                <Box sx={{ width: "90%", marginX: "5%" }}>
-                    <Box
-                        sx={{
-                            textAlign: "center",
-                            mb: 6
-                        }}
-                    >
-                        <SelectStatus
-                            id={props.id}
-                            status={props.status}
-                            setStatus={props.setStatus}
-                            responseStatus={props.responseStatus}
-                        />
-                    </Box>
-                    <Box
-                        sx={{
-                            display: "flex",
-                            flexDirection: "column"
-                        }}
-                    >
-                        <Box sx={{ display: "flex", alignItems: "center" }}>
-                            <Avatar
-                                alt="Student"
-                                src="/images/pose_english_shrug_man.png"
-                                sx={{
-                                    marginRight: 2
-                                }}
-                            />
-
-                            <Typography variant="h7" component="div">
-                                受講生 &nbsp; {props.updated_at}
-                            </Typography>
-                        </Box>
-
-                        <Box
-                            sx={{
-                                display: "flex",
-                                justifyContent: "space-between",
-                                alignItems: "center",
-                                mt: 2
-                            }}
-                        >
-                            <Box>
-                                <Typography variant="h7" component="div">
-                                    {categories[props.category]}&nbsp; /&nbsp;
-                                    {topics[props.topic]}
-                                    &nbsp; /&nbsp;
-                                    {props.curriculum_number}&nbsp;
-                                </Typography>
-                            </Box>
-                            <Box>
-                                <Link to={`/questions/` + props.id + `/edit`}>
-                                    <Button
-                                        variant="text"
-                                        sx={{ color: "#771AF8", fontSize: 15 }}
-                                    >
-                                        編集
-                                    </Button>
-                                </Link>
-                                /
-                                <Button
-                                    variant="text"
-                                    sx={{ color: "#771AF8", fontSize: 15 }}
-                                    onClick={props.deleteConfirm}
-                                >
-                                    削除
-                                </Button>
-                            </Box>
-                        </Box>
-                    </Box>
-
-                    <Box sx={{ display: "fex", position: "relative" }}>
-                        <Typography
-                            variant="h4"
-                            component="div"
-                            sx={{
-                                marginTop: 4,
-                                fontSize: 20,
-                                fontWeight: "bold"
-                            }}
-                        >
-                            {props.title}
-                        </Typography>
-                    </Box>
-
-                    <Typography
-                        variant="h6"
-                        component="div"
-                        sx={{
-                            marginTop: 3,
-                            borderBottom: "1px solid gray",
-                            fontWeight: "bold",
-                            padding: 1,
-                            fontSize: 15
-                        }}
-                    >
-                        調べたこと
-                    </Typography>
-                    <Typography
-                        variant="h5"
-                        component="div"
-                        sx={{
-                            fontSize: 15,
-                            padding: 1,
-                            marginBottom: 2
-                        }}
-                    >
-                        {props.remarks}
-                    </Typography>
-                    <Typography
-                        variant="h6"
-                        component="div"
-                        sx={{
-                            marginTop: 4,
-                            borderBottom: "1px solid gray",
-                            fontWeight: "bold",
-                            padding: 1,
-                            fontSize: 15
-                        }}
-                    >
-                        試したこと、分からないこと
-                    </Typography>
-                    <Typography
-                        variant="h5"
-                        component="div"
-                        sx={{
-                            padding: 1,
-                            marginBottom: 2,
-                            fontSize: 15
-                        }}
-                    >
-                        {question}
-                    </Typography>
-                </Box>
+                <QuestionContentMobile
+                    updated_at={props.updated_at}
+                    category={props.category}
+                    topic={props.topic}
+                    curriculum_number={props.curriculum_number}
+                    title={props.title}
+                    id={props.id}
+                    status={props.status}
+                    setStatus={props.setStatus}
+                    responseStatus={props.responseStatus}
+                    remarks={props.remarks}
+                    question={question}
+                />
             )}
         </>
     );
